Add files via upload
code_appendice.pdf 
This .pdf contains the links of the codes that have been borrowed for this presentation. 

thesis_document.pdf
This is the main written part of the thesis project, containing the literature review, methodology relevant research, implementation and testing and for the final part analysis of results. This document contains around 18,000 words all related to the project of keylogger detecting and the vulnerabilities of Microsoft Defender.

thesis_presentation.pptx
A thesis, would not be a thesis without your work explained in your own words - especially when you must explain it not only to two lecturers but your whole class containing a total of 28 students. This PowerPoint presentation contains the explained points in the document as well as videos which showcase the activity of keylogger, keylogger bypassing Microsoft Defender and the developed Python application detecting the keylogger - when Microsoft Defender fails to do so.