Add files via upload
These are the main parts of the project. This is what has been developed or composed by myself through this thesis which was used to perform penetration testing as well as defending. 

.TXT Uploader to DropBox
Contains a small Python script which will upload a .txt file from a designated location every 30/60 seconds. Our code does that because when the keylogger is running it is constantly updating a .txt document in a specific location, meaning that our .txt file is uploaded every 30/60 seconds which within that small period of time could allow a hacker to capture passwords or sensitive information.

C++ Converted Keylogger
This is a keylogger converted from Python into C++ this snippet of code never made it into the final part of the thesis other than the testing part, as we were testing our keylogger against Microsoft Defender's detection rate - we wanted to see if a C++ keylogger will be harder to detect than a Python keylogger, the detection rate difference was minor but C++ was detected less than Python.

Keylogger Detector
This is the main application of thesis, when we have shown that Microsoft Defender is vulnerable and does not detect a Python keylogger which has been freshly converted into an .exe using application such as PyInstaller our Python 'anti-virus' comes into play, as during our thesis we were asked to 'solve a problem' the idea could have been your own, upon finding out the Microsoft Defender vulnerability I found my goal for the thesis - which was a temporary and basic fix for a keylogger detector.

Python Keylogger
In short, this folder contains the basic Python keylogger code - this code has been borrowed from David Bombal which is included in the appendix.

Virus Downloader
Of course it would incomplete to showcase the scenario from the middle, I wanted to make sure that I showcase a real-life scenario example of how a user might be infected by a keylogger, how the keylogger bypasses basic defense systems (like Microsoft Defender in this case) and how my Python application fishes out the keylogger. So for the infection method I have used a fake 'Esset Antivirus' installer, clearly it would not be a very convincing application but with a descent amount of work put into it, you could even trick individual who are more advanced in I.T.